2d kernel submit, more stable key press, uint8 kernel output

+++ Cedai_Engine/src/Cedai.h
@@ -28,7 +28,7 @@ private:
 	Interface interface;
 	Renderer renderer;
 
-	float *pixels;
+	uint8_t* pixels;
 	float view[4][4] = { 0 };
 	uint32_t inputs;
